requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.1

# Machine Learning
scikit-learn>=1.6.0
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Flask Extensions
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-dotenv==1.0.0

# Language Support
googletrans==4.0.0rc1
langdetect==1.0.9

# Security
email-validator==2.1.0
cryptography==42.0.0

# Development & Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0

# Production
gunicorn==21.2.0
gevent==24.2.1
